[RA2 CH03]-Broken link (#1189)
* [RA2 CH03]-Broken link 

Broken link due to "/ref_model" repeated twice. Can anyone guide me, is this pointing to the right section. I believe it is supposed to point to 5.1 for software profile description and 5.3 for hardware profile description right?

* [RA2 CH03]-Broken link

Broken link due to "/ref_model" repeated twice. Pointing to 5.1 for software profile description and 5.3 for hardware profile description